pyproject.toml
[tool.poetry]
name = "fantacalcio-py-main"
version = "0.1.0"
description = ""
authors = ["piopy"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.13.4"
pandas = "^2.3.1"
loguru = "^0.7.3"
tqdm = "^4.67.1"
requests = "^2.32.4"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
